cloud_function/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub>=2.14.1
google-cloud-vision==3.1.4